Migrate to go mod. Improve Docker packaging.
Migrate the project to use go mods for easier
building and dependencies management.

Remove from the repository all the external packages
which are now pinned in the go.sum file.

Build and package the Docker image through a single
Dockerfile. The base image is changed to Alpine for
lighter packaging.

Small fixes to README.